// front-end/src/app/models/usuario.model.ts
export class Usuario {
    nick: string;
    name: string;
    lastName: string;
    email: string;
    birthdate: Date;
    nationality: string;
    typeDocument: string;
    numerDocument: string;
    codePhone: string;
    numerPhone: string
    university: string;
    faculty: string;
    occupation: string;
    password: string;
    password2: string
}

export class Countrie {
    countries: [];
}
export class Doctype {
    doctypes: [];
}
export class Gender {
    genders: [];
}